+++ src/App.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import './App.css';
 import { Grid } from '@material-ui/core' ;
-/* import TelaLogin from './Components/TelaLogin/TelaLogin.js'; */
-/* import TelaRegister from './Components/TelaRegister/TelaRegister.js'; */
-// import TelaRegister from './Components/EsqueceuSenha/EsqueceuSenha.js';
-// import ModificarSenha from './Components/ModificarSenha/ModificarSenha.js';
-// import CadastrarMonitoria from './Components/CadastrarMonitoria/CadastrarMonitoria.js';
-// import EditarPerfil from './Components/EditarPerfil/EditarPerfil.js';
-import TelaFeed from './Components/TelaFeed/TelaFeed.js';
+//  import Login from './Components/Login/Login.js'; 
+// import SignUp from './Components/SignUP/SignUp.js';
+//  import ForgotPassword from './Components/ForgotPassword/ForgotPassword.js';
+// import ModifyPassword from './Components/ModifyPassword/ModifyPassword.js';
+ //import RegisterMonitoring from './Components/RegisterMonitoring/RegisterMonitoring.js';
+ import EditProfile from './Components/EditProfile/EditProfile.js';
+ //import ReedBack from './Components/Feedback/Feedback.js';
  
 
 class App extends Component {
@@ -15,7 +15,7 @@ class App extends Component {
     return (
       <body className="BodyPrincipal">
       <Grid container  justify="center" >
-          <TelaFeed/>
+          <EditProfile/>
       </Grid>
       </body>
     );
