+++ src/Components/Feed/paper.js
