+++ src/Components/Feedback/Feedback.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import { Grid, Button, TextField } from '@material-ui/core' ;
-import BottomBar from '../BottomBar/BottomBar.js';
+import BottomBar from '../Tab/Tab.js';
 import AppBar from '../AppBar/AppBar.js';
 import Pp from '../../Assets/img/Pp.png';
 import Mencao from './Mencao.js';
 
 
 
-class AvaliarMonitor extends Component {
+class Feedback extends Component {
     
   render() {
       
@@ -54,4 +54,4 @@ class AvaliarMonitor extends Component {
   }
 }
 
-export default AvaliarMonitor;
+export default Feedback;
