+++ src/Components/Login/Login.css
@@ -1,0 +1,7 @@
+.LoginBackground{
+    position: fixed;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(180deg, #1DA1F2 65.09%, #1DA1F2 65.1%, #0E4C72 99.99%, #0E4C72 100%);
+
+}
