+++ src/Components/ModifyPassword/ModifyPassword.js
@@ -1,50 +1,54 @@
 import React, { Component } from 'react';
-import './TelaLogin.css';
 import { Grid, Button, TextField } from '@material-ui/core' ;
 import logo from '../../Assets/img/Logo.png';
+import './ModifyPassword.css';
 
 
 
-
-
-class TelaLogin extends Component {
-  
+class ModificarSenha extends Component {
   render() {
     return (
-      <div>
+      <div className="ModifyPasswordBackground">
         <Grid container alignContent="center" justify="center" direction="column" alignItems="center">
           <img src={logo} alt="Logo" />
           
           <Grid item >
             <TextField
-              id="emailTextField"
-              label="Email"
+              id="codigoTextField"
+              label="Código:"
               margin="normal"
               />
           </Grid>
           <Grid item >
             <TextField
-              id="senhaTextField"
-              label="Senha"
+              id="novaSenhaTextField"
+              label="Nova Senha"
               margin="normal"
               type="password"
              
               />
           </Grid>
-        </Grid>
-          <Grid container alignContent="center" justify="center" direction="column" spacing="24" alignItems="center" style={{marginTop: 25}}>
-            <Grid item >
-              <Button  variant="outlined">
-                  Login
-                </Button>
-              </Grid>
-              <Grid item>
-                <Button  variant="outlined">
-                  Registrar
-                </Button>
-              </Grid>
-              <a href="#">Esqueceu sua senha ?</a>  
+          <Grid item >
+            <TextField
+              id="confirmarSenhaTextField"
+              label="Confirmar nova Senha"
+              margin="normal"
+              type="password"
+              />
           </Grid>
+          </Grid>
+          <Grid container alignContent="center" justify="center" direction="column" spacing="24" alignItems="center" style={{marginTop: 25}}>
+          <Grid item >
+              <Button variant="outlined">
+                Enviar
+              </Button>
+              </Grid>
+            <Grid item>
+              <Button variant="outlined">
+                Voltar
+              </Button>
+            </Grid>
+            </Grid>
             
             
           
@@ -53,4 +57,4 @@ class TelaLogin extends Component {
   }
 }
 
-export default TelaLogin;
+export default ModificarSenha;
